Label config for bundles instead of annotating manifest
Some registries (gcr) are strict about following the docker manifest v2 schema
which does not allow annotations on a manifest while some others (artifactory)
do not support the OCI image manifest v1 spec. Because of this, we had to move
the marker of a bundle to the config, which all registries support.

Authored-by: Eli Wrenn <ewrenn@pivotal.io>
Signed-off-by: Eli Wrenn <ewrenn@pivotal.io>