The fix is simple and in bundle_images_lock
Added some tests for ImagesLock.LocalizeImagesLock
Removed the need for a pointer in the function since we no longer update
that reference

The tests for non distributable layers are failing.
This happens because the fake registry was indexing the images by
registry+repo and was not returning an error when a particular image was
not present in the registry.
In order to fix the problem, the FakeRegistry now references images by
full name+digest.
When the image does not exist in the registry instead of a dummy result
it returns an error