Ensure when collecting images the bundle repository is used
When collecting all the images that are part of a bundle, instead of
using `ImageRef.Image` to check if an image is a bundle, instead
Generate the locations and use the `ImageRef.Locations` to find where
the image can be found before trying to check.
This should provide 2 locations for each image, the Bundle Repository
and original Image.

- Major change on the `bundle_images_lock_test` to support an easier way
  to create a bundle and do assertions