Fallback to Get when Head Fails
In this commit we are falling back to a Get Request when the Head
request done to retrieve the Digest fails. This is done to prevent some
issues that we found with some registries when doing a HEAD Request,
where they were returning a 401 even for public images.
We decided to keep the Head request, because it is more performant and
will be sucessful in the majority of the cases.
This also helps due to some rate limitation that can happen when doing
Gets