Change the progress bar construction
go-containerregistry assumes that the channel that is provided to do
updates on the progress is a 1 use only channel. Previously we were
assuming that we could use the same channel for all requests.

This commit extracts the progress bar logger into a new Registry that
will be responsible for starting and stoping the progress bar when
MultiWrite is being called.

We assume that the progress bar will only be used for the MultiWrite, if
this is not the case in the future we can add similar logic to other
methods