Prefactor to include ImagesRef on Bundle struct
This prefactor will add all the Images that are associated to the
current bundle to a map. This will include all the images from the
current bundle plus all the images from the nested bundles.

This information will be used in next commits to distinguish between
what which images are bundles and which ones are not