bug fix: prevent null images in OCI location
When running imgpkg copy twice in a row, the second time would create an
OCI location file with the 'image' field set to null.

This was due to imgpkg 'detecting' that the image to copy was already
relocated with the bundle, and would update its image ref accordingly.

However when checking to see which images should be 'stored' into the
OCI location file, the image ref would not match the bundle's image ref
(it still had the original source for that image).

This fix checks all the 'locations' of the bundle's image ref, and if
any of them match (i.e. the location that the image was copied from)
then add the image ref to oci location.

Authored-by: Dennis Leon <leonde@vmware.com>