bug fix: prevent null images in OCI location
This occurs when a bundle contains an image reference missing the repo
name. i.e. 'ubuntu' is shorthand for 'library/ubuntu' on dockerhub.

when imgpkg decides whether the image should be included into the oci
location, it checks the unprocessed image ref (that will always contain
a repo) with the ref in the images lock (that might not contain the
repo)

This fix will *always* use the fully qualified name when returning all
the image refs for a bundle.

Authored-by: Dennis Leon <leonde@vmware.com>