UX: Return error if unable to determine which bundle to use in lockfile
This handles backwards compatibility. i.e. a user has generated a
tarball with an older version of imgpkg (that doesn't label which image
is the 'root bundle') and then uses a newer version of imgpkg that
expects to find which of the images is the 'root' bundle based on the
label.

Provides an error with a hint to re-generate the tarball if that is the case.

Authored-by: Dennis Leon <leonde@vmware.com>