Assert there are no file changes when building imgpkg
- attempts to handle the fact we have modified a vendor file and apply a
patch to disable the default docker provider. If, say dependabot bumps
that package, and we forget to commit the change into the vendor
directory, the patch will apply successfully, leaving unstaged git
changes. This check will cause the build to fail - acting as a reminder
to commit the patch changes back in.

Authored-by: Dennis Leon <leonde@vmware.com>