Ensure image is mountable
Checks if an image can be read before deciding if it is mountable. The
scenario more common for errors is having 2 repositories that 2
different users can access. When copying the Image `imgpkg` needs to
make sure that the destination user can read the source in order to
mount the Image. If they do not have, it should push the layer.