Remove --output-type option from pull
This commit can be reverted in order to get this feature back to the
code. We decided to, for now, remove it since we are not going to use it
in the short term.
Nevertheless the basic implementation will still be present in case we
want to reintroduce this change in the future.

Signed-off-by: Joao Pereira <joaod@vmware.com>