Copy with resume reuses the file instead of copying
With this change we are going to use the same file that was provided to
be resumed. We do this to increase the performance, previously we were
copying the file to the side, which sounds more safe but the cost of the
copy vs move is too high.
This change was able to shave almost 20 minutes from a copy that was
taking around 1h10.

Signed-off-by: João Pereira <joao.pereira@broadcom.com>